Rename propTypes
